rename files, add more compose env files, prepare saved_searches
