+++ app/deploy/nginx-dev.conf
@@ -1,0 +1,28 @@
+server_tokens off;
+
+server {
+       listen 80;
+       server_name nginx;
+       charset utf-8;
+
+       access_log /var/log/nginx/app-access.log;
+       error_log /var/log/nginx/app-error.log;
+
+       # max upload size
+       client_max_body_size 4096M; # adjust to taste
+
+       # phpmyadmin
+       location /phpmyadmin/ {
+            proxy_pass http://phpmyadmin/;
+            #proxy_set_header Host $http_host;
+            #proxy_set_header X-Forwarded-For $proxy_add_x_forwarded_for;
+            #proxy_set_header X-Forwarded-Server $http_host;
+            #proxy_redirect / /phpmyadmin/;
+            #proxy_cookie_path / /phpmyadmin/;
+       }
+
+       location / {
+	     uwsgi_pass app:8000;
+	     include /etc/nginx/uwsgi_params;
+       }
+}
